fixing classpath issues
